feat: move dir for tracing
